requirements.txt
sqlalchemy==2.0.36
typer==0.15.1
rich==13.9.4
httpx==0.28.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
